Add config option for JWT issuer and require iat/exp claims in token
Runtime users should be validating the JWT issuer in addition to the
JWKS. This commit adds logic to configure the expected issuer for a
JWT and updates the token parser to require the "iat" and "exp" claims
be present for a given token. Additionally, this commit fixes an error
where command line flags were not being bound to Viper configs
properly.

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>