Add support for publishing events to iam-runtime-infratographer
This commit adds support for CreateRelationships and
DeleteRelationships in the runtime implementation. Unlike the general
events package, only a single topic is configured here in the config
file to make general administration of NATS easier.

This commit also updates the README to include NATS configuration
options and context about managing relationships.

Signed-off-by: John N Schaeffer <jschaeffer@equinix.com>