implement iam-runtime GetAccessToken
This implements the GetAccessToken method which returns an access token upon request.

Returned tokens can be configured to be from a file or client credentials.
Additionally the sourced token can be exchanged for another token.

This is particularly useful within a kubernetes cluster.
You can exchange the service account token for one from another issuer.

Signed-off-by: Mike Mason <mimason@equinix.com>