provide failure message when not using the new manifests helper
With the split of the config into both a configmap and secrets a new helper
was created `manifests` which will include both of these manifests.

This however means that the configmap helper must no longer be used otherwise
the secrets manifest resulting in missing dependencies.

To ensure this breaking change is clear, we renamed configmap and put a deprecation
message in it's place, informing the user of how to resolve the issue.

Signed-off-by: Mike Mason <mimason@equinix.com>